requirements.txt
streamlit>=1.37,<1.39
pandas>=2.2.3
numpy>=2.0,<3
plotly>=5.23,<6
psycopg2-binary>=2.9
python-dotenv>=1.0
google-generativeai>=0.7
streamlit-autorefresh>=1.0
